+++ desktop-widgets/tagwidget.h
@@ -26,7 +26,11 @@ slots:
 	void completionHighlighted(const QString &text);
 
 protected:
-	void keyPressEvent(QKeyEvent *e);
+	void keyPressEvent(QKeyEvent *e) override;
+	void dragEnterEvent(QDragEnterEvent *e) override;
+	void dragLeaveEvent(QDragLeaveEvent *e) override;
+	void dragMoveEvent(QDragMoveEvent *e) override;
+	void dropEvent(QDropEvent *e) override;
 private:
 	void focusOutEvent(QFocusEvent *ev) override;
 	QCompleter *m_completer;